requirements.txt
# Veri işleme
pandas==2.2.2
numpy==1.26.4

# Makine öğrenmesi
scikit-learn==1.5.1
xgboost==2.1.1
lightgbm==4.5.0
imbalanced-learn==0.12.3

# Görselleştirme
matplotlib==3.9.2
seaborn==0.13.2

# Notebook
jupyter==1.1.1
ipykernel==6.29.5

# Yardımcı
joblib==1.4.2
tqdm==4.66.5
pyyaml==6.0.2
